+++ src/chromyRunner.test.js
@@ -2,7 +2,8 @@ import fs from 'fs';
 import Chromy from 'chromy';
 import { ChromyRunner } from './chromyRunner';
 import logger from './logger';
-import { configTypes, testReportStep } from './defaultConfig';
+import { globalConfig, testConfig, configTypes } from './defaultConfig';
+import actions from './actions';
 
 jest.mock('chromy', () => jest.fn().mockImplementation(() =>
     ({
@@ -15,7 +16,7 @@ jest.mock('chromy', () => jest.fn().mockImplementation(() =>
 
 jest.mock('./compareImage', () => jest.fn(arg =>
     new Promise((resolve, reject) => {
-      if (arg === './screenshots/default.png') {
+      if (arg === './screenshots') {
         return resolve('Writting the diff image to disk');
       }
       return reject('error');
@@ -30,25 +31,6 @@ logger.log = (...args) => {
 let writeFileSyncCalls = [];
 fs.writeFileSync = (...args) => {
   writeFileSyncCalls.push(...args);
-};
-
-const globalConfig = {
-  screenshots: './screenshots',
-  debug: false,
-  visible: true,
-  timeout: 30000,
-};
-
-const testConfig = {
-  name: 'default',
-  resolution: {
-    width: 800,
-    height: 600,
-  },
-  steps: [
-    { name: 'goto', value: 'www.example.com' },
-    { name: 'capture', value: 'document' },
-  ],
 };
 
 const chromyRunner = new ChromyRunner(globalConfig);
@@ -67,10 +49,12 @@ describe('ChromyRunner', () => {
     expect(loggerCalls[1]).toEqual('screenshot saved in -> ./screenshots/default.png');
     expect(chromyRunner.options)
       .toEqual({
-        debug: false,
         screenshots: './screenshots',
-        timeout: 30000,
+        testReportPath: './differencify_report',
+        debug: false,
         visible: true,
+        timeout: 30000,
+        mismatchThreshold: 0.1,
       });
     expect(writeFileSyncCalls).toEqual(['./screenshots/default.png', 'png file']);
   });
@@ -80,28 +64,30 @@ describe('ChromyRunner', () => {
     expect(result).toEqual(true);
     expect(chromyRunner.currentTestId).toEqual(2);
     expect(loggerCalls[0]).toEqual('goto -> www.example.com');
-    expect(loggerCalls[1]).toEqual('screenshot saved in -> differencify_report/default.png');
+    expect(loggerCalls[1]).toEqual('screenshot saved in -> ./differencify_report/default.png');
     expect(chromyRunner.options)
       .toEqual({
-        debug: false,
         screenshots: './screenshots',
-        timeout: 30000,
+        testReportPath: './differencify_report',
+        debug: false,
         visible: true,
+        timeout: 30000,
+        mismatchThreshold: 0.1,
       });
-    expect(writeFileSyncCalls).toEqual(['differencify_report/default.png', 'png file']);
+    expect(writeFileSyncCalls).toEqual(['./differencify_report/default.png', 'png file']);
   });
   describe('Chromy runner', () => {
     it('Step runner: test action', async () => {
       const chromy = new Chromy();
       testConfig.type = configTypes.test;
-      testConfig.steps.push(testReportStep);
+      testConfig.steps.push({ name: actions.test, value: globalConfig.testReportPath });
       const result = await chromyRunner._stepRunner(chromy, testConfig);
-      testConfig.steps.pop(testReportStep);
+      testConfig.steps.pop({ name: actions.test, value: globalConfig.testReportPath });
       expect(result).toEqual(true);
       expect(loggerCalls[0]).toEqual('goto -> www.example.com');
-      expect(loggerCalls[1]).toEqual('screenshot saved in -> differencify_report/default.png');
+      expect(loggerCalls[1]).toEqual('screenshot saved in -> ./differencify_report/default.png');
       expect(loggerCalls[2]).toEqual('Writting the diff image to disk');
-      expect(writeFileSyncCalls).toEqual(['differencify_report/default.png', 'png file']);
+      expect(writeFileSyncCalls).toEqual(['./differencify_report/default.png', 'png file']);
     });
     it('Step runner: update action', async () => {
       const chromy = new Chromy();