+++ src/compareImage.js
@@ -1,19 +1,20 @@
 
 import Jimp from 'jimp';
 import logger from './logger';
-import { testReportStep } from './defaultConfig';
 
-const compareImage = (referencePath, testPath, misMatchThreshold) =>
+const compareImage = (referencePath, testPath, misMatchThreshold, testName) =>
   new Promise((resolve, reject) => {
-    Jimp.read(referencePath).then((referenceImage) => {
-      Jimp.read(testPath).then((testImage) => {
+    const referenceFile = `${referencePath}/${testName}.png`;
+    const testFile = `${testPath}/${testName}.png`;
+    Jimp.read(referenceFile).then((referenceImage) => {
+      Jimp.read(testFile).then((testImage) => {
         const distance = Jimp.distance(referenceImage, testImage);
         const diff = Jimp.diff(referenceImage, testImage, misMatchThreshold);
         logger.log('Writting the diff image to disk');
-        diff.image.write(`${testReportStep.value}/differencified.png`);
         if (distance < misMatchThreshold || diff.percent < misMatchThreshold) {
           return resolve('No mismatch found!');
         }
+        diff.image.write(`${testPath}/${testName}_differencified.png`);
         return reject('Mismatch found!');
       }).catch((err) => {
         logger.error(err);