+++ src/index.test.js
@@ -13,7 +13,7 @@ jest.mock('chromy', () => jest.fn().mockImplementation(() =>
 
 jest.mock('./compareImage', () => jest.fn(arg =>
     new Promise((resolve, reject) => {
-      if (arg === 'screenshots/default.png') {
+      if (arg === 'screenshots') {
         return resolve('Writting the diff image to disk');
       }
       return reject('error');
@@ -64,14 +64,14 @@ describe('Differencify', () => {
     expect(result).toEqual(true);
     expect(loggerCalls[0]).toEqual('goto -> www.example.com');
     expect(loggerCalls[1]).toEqual('screenshot saved in -> screenshots/default.png');
-    expect(writeFileSyncCalls).toEqual(['screenshots', 'differencify_report', 'screenshots/default.png', 'png file']);
+    expect(writeFileSyncCalls).toEqual(['screenshots', './differencify_report', 'screenshots/default.png', 'png file']);
   });
   it('test fn', async () => {
     const result = await differencify.test(testConfig);
     expect(result).toEqual(true);
     expect(loggerCalls[0]).toEqual('goto -> www.example.com');
-    expect(loggerCalls[1]).toEqual('screenshot saved in -> differencify_report/default.png');
+    expect(loggerCalls[1]).toEqual('screenshot saved in -> ./differencify_report/default.png');
     expect(loggerCalls[2]).toEqual('Writting the diff image to disk');
-    expect(writeFileSyncCalls).toEqual(['differencify_report/default.png', 'png file']);
+    expect(writeFileSyncCalls).toEqual(['./differencify_report/default.png', 'png file']);
   });
 });